Добавил проверку наличия сущности в бд при ее создании
В use case CreateEventUseCase добавил проверку наличия сущности в бд при ее создании

+++ Application/Modsen.Application/Use Case/CreateEventUseCase.cs
@@ -4,33 +4,39 @@ using Modsen.Domain;
 
 namespace Modsen.Application
 {
-    public class CreateEventUseCase
+public class CreateEventUseCase
+{
+    private readonly IEventRepository _eventRepository;
+    private readonly ImageService _imageService;
+
+    public CreateEventUseCase(IEventRepository eventRepository, ImageService imageService)
     {
-        private readonly IEventRepository _eventRepository;
-        private readonly ImageService _imageService;
-
-        public CreateEventUseCase(IEventRepository eventRepository, ImageService imageService)
-        {
-            _eventRepository = eventRepository;
-            _imageService = imageService;
-        }
-
-        public async Task<MyEvent> CreateEventAsync(
-            MyEvent newEvent, 
-            List<IFormFile> eventImages, 
-            IWebHostEnvironment hostEnvironment, 
-            CancellationToken cancellationToken)
-        {
-            newEvent.EventImages = _imageService.SaveImages(eventImages, hostEnvironment);
-
-            await _eventRepository.AddEventAsync(newEvent, cancellationToken);
-
-            await _eventRepository.SaveChangesAsync(cancellationToken);
-
-            _imageService.DeleteUnusedImages(hostEnvironment, _eventRepository, cancellationToken);
-
-            return newEvent;
-        }
+        _eventRepository = eventRepository;
+        _imageService = imageService;
     }
+
+    public async Task<MyEvent> CreateEventAsync(
+        MyEvent newEvent, 
+        List<IFormFile> eventImages, 
+        IWebHostEnvironment hostEnvironment, 
+        CancellationToken cancellationToken)
+    {
+        var existingEvent = await _eventRepository.GetEventByNameAsync(newEvent.Name, cancellationToken);
+        if (existingEvent != null)
+        {
+            throw new BadRequestException($"Event with name '{newEvent.Name}' already exists.");
+        }
+
+        newEvent.EventImages = _imageService.SaveImages(eventImages, hostEnvironment);
+
+        await _eventRepository.AddEventAsync(newEvent, cancellationToken);
+        await _eventRepository.SaveChangesAsync(cancellationToken);
+
+        _imageService.DeleteUnusedImages(hostEnvironment, _eventRepository, cancellationToken);
+
+        return newEvent;
+    }
+}
+
 
 }
